+++ source/extensions/lightspeed.trex.stage_manager.plugin.interaction.usd/lightspeed/trex/stage_manager/plugin/interaction/usd/all_meshes.py
@@ -52,6 +52,7 @@ class AllMeshesInteractionPlugin(_StageManagerUSDInteractionPlugin):
         default=[
             "AssignCategoryActionWidgetPlugin",
             "CustomTagsWidgetPlugin",
+            "DeleteRestoreActionWidgetPlugin",
             "FocusInViewportActionWidgetPlugin",
             "IsCaptureStateWidgetPlugin",
             "IsCategoryHiddenStateWidgetPlugin",